Update FlaGs_linux.py
As per Biopython documentation: "Bio.Alphabet has been removed from Biopython. In many cases, the alphabet can simply be ignored and removed from scripts."
Also, the install script needs to be updated so as not to install 'ete_toolchain' in the conda environment (it's a dead project and the new 'ete3' package seems to be working fine from my end, but this may need to be verified).